Organiza APIs e database em pastas e arquivos separados
