+++ exercicios-obrigatorio/database/todolists_db.js
@@ -1,0 +1,16 @@
+module.exports = {
+    toDoLists: [
+        {
+            title: 'title1',
+            text: 'text...'
+        },
+        {
+            title: 'title2',
+            text: 'text...'
+        },
+        {
+            title: 'title3',
+            text: 'text...'
+        }
+    ]
+}
