+++ exercicios-obrigatorio/database/users_db.js
@@ -1,0 +1,19 @@
+module.exports = {
+    users: [
+        {
+            name: 'fulano',
+            email: 'fulano@hotmail.com',
+            password: 'senha1'
+        },
+        {
+            name: 'beltrano',
+            email: 'beltrano@hotmail.com',
+            password: 'senha2'
+        },
+        {
+            name: 'cicrano',
+            email: 'cicrano@hotmail.com',
+            password: 'senha3'
+        }
+    ]
+}
